+++ dao/src/main/kotlin/fund/cyber/markets/cassandra/repository/TickerRepository.kt
@@ -8,10 +8,10 @@ import fund.cyber.markets.cassandra.MARKETS_KEYSPACE
 import fund.cyber.markets.cassandra.PREFERRED_CONCURRENT_REQUEST_TO_SAVE_ENTITIES_LIST
 import fund.cyber.markets.cassandra.accessor.TickerAccessor
 import fund.cyber.markets.common.Durations
+import fund.cyber.markets.common.closestSmallerMultiply
 import fund.cyber.markets.dto.TokensPair
 import fund.cyber.markets.helpers.addHop
 import fund.cyber.markets.model.Ticker
-import fund.cyber.markets.util.closestSmallerMultiply
 import io.reactivex.Flowable
 import java.util.*
